build: Refactor Home into aQute.bnd.util.home package
Some bundles -conditionalpackage aQute.bnd.util.* so we place Home under
the aQute.bnd.util package tree.
