+++ biz.aQute.bndlib/src/aQute/bnd/http/HttpClient.java
@@ -45,7 +45,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import aQute.bnd.connection.settings.ConnectionSettings;
-import aQute.bnd.home.Home;
 import aQute.bnd.http.URLCache.Info;
 import aQute.bnd.osgi.Processor;
 import aQute.bnd.service.Registry;
@@ -57,6 +56,7 @@ import aQute.bnd.service.url.State;
 import aQute.bnd.service.url.TaggedData;
 import aQute.bnd.service.url.URLConnectionHandler;
 import aQute.bnd.service.url.URLConnector;
+import aQute.bnd.util.home.Home;
 import aQute.lib.io.IO;
 import aQute.lib.json.JSONCodec;
 import aQute.libg.reporter.ReporterAdapter;
