+++ biz.aQute.bndlib/src/aQute/bnd/maven/MavenCommand.java
@@ -31,7 +31,6 @@ import org.slf4j.LoggerFactory;
 import aQute.bnd.header.Attrs;
 import aQute.bnd.header.OSGiHeader;
 import aQute.bnd.header.Parameters;
-import aQute.bnd.home.Home;
 import aQute.bnd.maven.support.CachedPom;
 import aQute.bnd.maven.support.Maven;
 import aQute.bnd.maven.support.Pom;
@@ -42,6 +41,7 @@ import aQute.bnd.osgi.Descriptors.PackageRef;
 import aQute.bnd.osgi.Jar;
 import aQute.bnd.osgi.Processor;
 import aQute.bnd.osgi.Resource;
+import aQute.bnd.util.home.Home;
 import aQute.lib.collections.LineCollection;
 import aQute.lib.io.IO;
 import aQute.lib.settings.Settings;
